+++ api/models.py
@@ -1,3 +1,4 @@
+import typing
 from pydantic import BaseModel, AnyUrl, validator
 
 
@@ -7,8 +8,8 @@ class EventGeneratorRequest(BaseModel):
     event_type: str
     event_subject: str
     event_data: str
-    iterations: str
-    delay: str | None
+    iterations: int = 1
+    delay: int = 100
 
     @validator('iterations')
     def check_positive_iterations(cls, value):
@@ -26,3 +27,13 @@ class EventGeneratorRequest(BaseModel):
 
     class Config:
         extra = 'forbid'
+
+
+class EventGeneratorTask(BaseModel):
+    id: str
+    status: str
+    progress: int = 0
+    client_id: str | None
+
+    def __setitem__(self, key: str, value: typing.Any):
+        setattr(self, key, value)